NewMethods
On "ConsultarMedicamentoPeloNome" file, created two methods; one for asking for the name of the medicine and the other for looking for it on the csv file.

On "Main" created a Method for the Menu Interface and another for the chioce as well as another for handling said choice.

On "ReadCSV.java", under Testes, improved content visualization and numbering of rows and columns